# sql: avoid net.Error allocation on each readTimeoutConn.Read call